Add example from page 8
In the paper on page 8, this is an example method called `exp`. In a
larger code base like this one, I found `Exp` too generic a name, so
following a prior example, I named it `Make3Plus4`.

I also added the `test` method, but again had to rename it.

It's unclear to me which `eval` method this line of code refers to:

    int x = exp(base).eval();

Searching through the paper's text prior to page 8, the only `eval`
method in scope is the method defined by the `Exp` interface in figure
1. This method, however, returns a `Value`, and not an `int`. As far as
I can tell, the above line of code might have worked if an additional
`getInt()` had been appended:

    int x = exp(base).eval().getInt();

Am I missing something, or is this an error in the paper?

In any case, I've replaced that line with a call to `Accept`. The result
is equivalent, so I don't think that it makes much of a difference.