Revert "Delete redundant IValue API"
This reverts commit 3852843498250fdd2f8c824146ca6e1470f8f46a.

After I ran into the problem of how to infer the 'current' definition of
`IExp`, I posted a question on Stack Overflow:
https://stackoverflow.com/q/67818254/126014 After publishing the link on
Twitter (https://twitter.com/ploeh/status/1400428323717648385), I
received an answer from Bruno C. d. S. Oliveira, one of the authors of
the paper!

It turns out that the definition of `Exp` assumed in most of the paper
is the one that returns a `Value` (what I call `IValue`), so I'm
reverting this commit to bring back that interface and its
implementations.