Add Eval method to IExp
This continues the ongoing work of bringing back the `Eval` method to
`IExp`, as explained by Bruno C. d. S. Oliveira on Stack Overflow:
https://stackoverflow.com/a/67847476/126014.

I'm moving in small steps here, as this is a refactoring that may affect
so many moving parts (including tests) that I find it most prudent to
move in as small steps as possible. As Tim Ottinger writes: move by
micro-commits:
https://www.industriallogic.com/blog/whats-this-about-micro-commits

This is the reason I've yet to implement the `Eval` method on any of the
`IExp` implementations. This comes next.