Add StmtFactory
The class is only partially implemented, as I wanted to stop and add a
simple test of the interaction of `Assign` and `Var` before continuing
the implementation.

As I understand the implied 'language' modelled by StmtFactory, it uses
eager evaluation, since its map keeps track of evaluated values assigned
to variables, rather than expressions (thunks).

I do find it odd, however, that Assign only evaluates and adds the value
to the map when the returned expression is evaluated. You can see this
by changing the order of the two assertions in the `AssignLitVar` test.
In other words, evaluating the expression returned by `Assign` has the
side effect of 'registering' the variable. Even worse, you MUST evaluate
that expression before evaluating the `Var` expression!

This seems to me a bug, but is how figure 9 implements the interface,
unless I'm completely misuunderstanding how Java anonymous classes work.

Still, this doesn't seem like a dealbreaker. I think it'd be possible to
move the evaluation to the `Assign` method itself, so that it forces
evaluation. I'm not, however, going to rock that boat right now, since I
still don't haven't read the entire paper, and thus, I don't know what's
going to happen next.