Implement the rest of StmtFactory
I added no tests yet, since it's not clear to me how one is supposed to
interact with this API. How do we sense that IStmt.Eval is invoked?

I could, of course, add some sort of Test Spy, but is that the intent?

I'm getting the impression that this example 'language' is supposed to
be 'useful' in itself. How exactly that's supposed to happen still
eludes me.

Reading on, hoping to be illuminated.