Add Debug class
I had to make `Combine.Add` virtual. In Java (as the paper uses),
methods are virtual by default. In C#, you explicit have to make them
virtual if you want them to be so.

I could also have made the `Lit` method virtual. That would have been
more consistent, but isn't required for the example, and this is, after
all, throw-away code.

I also decided to override `VInt.ToString` because by default it only
returns the type name, and I didn't find that useful. Again, for
consistency, one might want to override `VBool.TosString` as well, but
again this wasn't required for the example.